moved database related functions to model layer
